// src/Models/PokemonDetails/SpriteUrls.ts
import { ISpritesUrls } from "./ISpritesUrls";

export class SpritesUrls implements ISpritesUrls {
    back_default?: string;
    back_female?: string;
    back_shiny?: string;
    back_shiny_female?: string;
    front_default?: string;
    front_female?: string;
    front_shiny?: string;
    front_shiny_female?: string;  

}